Refactor classes to use anakin's command result and commandexception
