+++ src/main/java/seedu/address/logic/anakincommands/EditCardCommand.java
@@ -12,8 +12,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.CommandHistory;
-import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.anakincommands.exceptions.CommandException;
 
 import seedu.address.model.Model;
 import seedu.address.model.anakindeck.Answer;
