+++ src/main/java/seedu/address/logic/anakincommands/SortCommand.java
@@ -3,7 +3,6 @@ package seedu.address.logic.anakincommands;
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.CommandHistory;
-import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 /**
