+++ src/test/java/seedu/address/logic/anakincommands/ExitCommandTest.java
@@ -9,7 +9,6 @@ import org.junit.Test;
 
 import seedu.address.commons.events.ui.ExitAppRequestEvent;
 import seedu.address.logic.CommandHistory;
-import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.ui.testutil.EventsCollectorRule;
